Factor out SshExec and report any command failures
